@@ -5,7 +5,7 @@
 import {Component, OnInit} from '@angular/core';
 import {ReleaseService} from '../release.service';
 import {CollectionViewer, DataSource} from '@angular/cdk/collections';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {MatDialog} from '@angular/material';
 import {NewReleaseDialogComponent} from '../new-release-dialog/new-release-dialog.component';
 import * as QuillDeltaToHtmlConverter from 'quill-delta-to-html';
