@@ -4,7 +4,7 @@
 
 import {Injectable} from '@angular/core';
 import {HttpClient, HttpHeaders} from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {Release, ReleaseFile} from './release/release.model';
 
 @Injectable()
