@@ -10,19 +10,18 @@ const Minio = require('minio');
 const config = require('../config/config.dev');
 
 const minioClient = new Minio.Client({
-  endPoint: config.objectStore.endpoint, 
+  endPoint: config.objectStore.endpoint,
   port: config.objectStore.port,
   secure: false,
-  accessKey: config.objectStore.accessKey, 
+  accessKey: config.objectStore.accessKey,
   secretKey: config.objectStore.secretKey
 });
 
 router.get('/file/:file', (req, res) => {
-  console.log('file/:file', req.params.file);
   minioClient.getObject('test-release', req.params.file, (error, stream) => {
     if(error){
       console.error(error);
-      res.status(500).send(error);
+      res.status(404).send(error);
       return;
     }
 
@@ -40,7 +39,6 @@ router.get('/latest/:platform', (req, res, next) => {
 });
 
 router.get('/release/:id/:fileID', (req, res) => {
-  console.log('/release/:id/:fileID');
   const id = req.params.id;
   const fileID = req.params.fileID;
   releaseController.getFile(id, fileID)
